# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge RAG API
  version: 1.2.0
  description: |
    基于当前代码实现整理的接口规范，覆盖登录鉴权、注册码管理、文档上传管理、公开文档浏览与 RAG 问答接口。

    说明：
    - 文档管理接口使用 Session Cookie 鉴权，登录成功后服务端会写入 `JSESSIONID`。
    - 公开文档接口（`/documents/public/**`）无需鉴权，所有限流保护。
    - RAG 问答接口保持匿名可访问。
    - `/auth/register` 只能使用管理员创建的一次性注册码注册。
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: 登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: admin
              password: ChangeMe123!
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
              example:
                message: 登录成功
                user:
                  username: admin
                  role: ADMIN
        '400':
          description: 参数不合法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Auth
      summary: 使用注册码注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: editor01
              password: Password123
              registrationCode: ABCD-EFGH-JKLM
      responses:
        '201':
          description: 注册成功，并自动登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
              example:
                message: 注册成功
                user:
                  username: editor01
                  role: USER
        '400':
          description: 用户名冲突、注册码无效、已使用、已禁用或已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/registration-codes:
    get:
      tags:
        - Auth
      summary: 管理员查看注册码列表
      operationId: listRegistrationCodes
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 注册码列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationCodeListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags:
        - Auth
      summary: 管理员创建注册码
      operationId: createRegistrationCode
      security:
        - cookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationCodeCreateRequest'
            example:
              note: 运营同事
              expiresAt: '2026-04-01T18:30:00'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationCodeResponse'
        '400':
          description: 到期时间不合法等参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /auth/registration-codes/{id}/disable:
    patch:
      tags:
        - Auth
      summary: 管理员禁用注册码
      operationId: disableRegistrationCode
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/RegistrationCodeId'
      responses:
        '200':
          description: 禁用成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationCodeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: 注册码不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/registration-codes/{id}:
    delete:
      tags:
        - Auth
      summary: 管理员删除注册码
      operationId: deleteRegistrationCode
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/RegistrationCodeId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: 注册码已删除
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: 注册码不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /documents/public/{documentId}:
    get:
      tags:
        - Public Documents
      summary: 获取公开文档详情
      operationId: getPublicDocumentDetail
      description: 返回文档元数据及全部文本段落，无需鉴权。受 IP 限流保护（30 次/分钟）。
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: 文档详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDocumentDetailResponse'
        '404':
          description: 文档不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 查询失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    UnauthorizedError:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 未登录
            message: 请先登录后再访问该接口
            timestamp: '2026-03-29T18:00:00'
    ForbiddenError:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 无权限
            message: 当前账号无权执行该操作
            timestamp: '2026-03-29T18:00:00'
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      description: 文档唯一标识
      schema:
        type: string
...
    RegistrationCodeId:
      name: id
      in: path
      required: true
      description: 注册码主键 ID
      schema:
        type: integer
        format: int64
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - registrationCode
      properties:
        username:
          type: string
        password:
          type: string
        registrationCode:
          type: string
    AuthUserResponse:
      type: object
      required:
        - username
        - role
      properties:
        username:
          type: string
        role:
          type: string
          example: ADMIN
    AuthSuccessResponse:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/AuthUserResponse'
...
    RegistrationCodeCreateRequest:
      type: object
      properties:
        note:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
    RegistrationCodeListResponse:
      type: object
      required:
        - codes
        - total
      properties:
        codes:
          type: array
          items:
            $ref: '#/components/schemas/RegistrationCodeResponse'
        total:
          type: integer
          format: int32
    RegistrationCodeResponse:
      type: object
      required:
        - id
        - code
        - createdBy
        - createdAt
        - status
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        note:
          type: string
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        usedAt:
          type: string
          format: date-time
          nullable: true
        usedBy:
          type: string
          nullable: true
        disabledAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [AVAILABLE, USED, DISABLED, EXPIRED]
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    PublicDocumentSegment:
      type: object
      required:
        - chunkIndex
        - text
      properties:
        chunkIndex:
          type: integer
          format: int32
        text:
          type: string
...
    PublicDocumentDetailResponse:
      type: object
      required:
        - documentId
        - filename
        - title
        - category
        - documentTime
        - keywords
        - segmentCount
        - segments
      properties:
        documentId:
          type: string
        filename:
          type: string
        title:
          type: string
        category:
          type: string
        documentTime:
          type: string
        keywords:
          type: string
        segmentCount:
          type: integer
          format: int32
          minimum: 0
        segments:
          type: array
          items:
            $ref: '#/components/schemas/PublicDocumentSegment'
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...